Move Condition and Effect classes to separate module.
